Moved router to a new a file and added config file
